+++ CHANGES.md
@@ -5,6 +5,10 @@ This file records the changes in each m3api-oauth2 release.
 The annotated tag (and GitHub release) for each version also lists the changes,
 but this file may sometimes contain later improvements (e.g. typo fixes).
 
+## next (not yet released)
+
+No changes yet.
+
 ## v0.3.1 (2024-09-07)
 
 - Updated m3api requirement to the latest version (0.8.3) –